langchain-vertex-test/hello-world/requirements.txt
langchain-google-genai
pillow
python-dotenv
